feat: auth 설정 관련 로직 분리
