+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,22 +1,5 @@
-import NextAuth, { AuthOptions } from "next-auth";
-import GoogleProvider from "next-auth/providers/google";
-
-function getEnvVariable(key: string): string {
-  const value = process.env[key];
-  if (!value) {
-    throw new Error(`Missing environment variable: ${key}`);
-  }
-  return value;
-}
-
-export const authOptions: AuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: getEnvVariable("GOOGLE_CLIENT_ID"),
-      clientSecret: getEnvVariable("GOOGLE_CLIENT_SECRET"),
-    }),
-  ],
-};
+import { authOptions } from "@/lib/auth";
+import NextAuth from "next-auth";
 
 const handler = NextAuth(authOptions);
