+++ src/lib/auth.ts
@@ -1,0 +1,21 @@
+import NextAuth, { AuthOptions } from "next-auth";
+import GoogleProvider from "next-auth/providers/google";
+
+function getEnvVariable(key: string): string {
+  const value = process.env[key];
+  if (!value) {
+    throw new Error(`Missing environment variable: ${key}`);
+  }
+  return value;
+}
+
+export const authOptions: AuthOptions = {
+  providers: [
+    GoogleProvider({
+      clientId: getEnvVariable("GOOGLE_CLIENT_ID"),
+      clientSecret: getEnvVariable("GOOGLE_CLIENT_SECRET"),
+    }),
+  ],
+};
+
+export default NextAuth(authOptions);
